Prevent index from failing silently
There is a maximum size for program arguments (see ARG_MAX), and
xargs(1) automatically invokes the utility multiple times if this
limit would be exceeded.

In this case that would silently generate a partial index, so this
change uses command expansion in the shell which returns an error
if the arguments are too large.